\name{dat2i}
\alias{dat2i}
\docType{data}

\title{Information for httk function suggets.httk()
}

\description{A lsit with information needed for httk.fun()}

\usage{data("dat2i")}

\format{
A lsit of two objects
 \item{parameters}{A lsit with all possible parameters used to bulit the model in httk.fun()}
 \item{predictedFeatures}{Variable names from httk.fun() raw model}
}

\details{Example dataset to construct the resulting data matrix drom suggest.httk}

\source{A toy example for httk.fun based on jaqpot specifications}

\references{httk cran package}

\examples{
data(dat2i)
## maybe str(dat2i) ; plot(dat2i) ...
}

\keyword{datasets2i}